refactor InvocationTranspiler out of HandlebarsPreprocessor (#200)
This commit refactors the new InvocationTranspiler class out of
the HandlebarsPreprocessor. This work ended up not being strictly
needed for using the Handlebars parser to parse out translate helper
usages, but I still thought it was a nice tech debt change.
No new functionality was added, just a number of methods were split
off into InvocationTranspiler.

J=SLAP-1178
TEST=manual

test building a site with translation helpers
also see that multiline translation helpers still fail to build
